refactor: standardize variable naming and simplify logic for tag component
