+++ src/pages/EventView.tsx
@@ -1,4 +1,4 @@
-import CustomTag from "@/components/buttons/CustomTag/CustomTag";
+import Tag from "@/components/buttons/tag/tag";
 import FlowButton from "@/components/buttons/FlowButton/FlowButton";
 import PrimaryCustomButton from "@/components/buttons/PrimaryCustomButton/PrimaryCustomButton";
 import Frame from "@/components/layout/frame";
@@ -27,10 +27,10 @@ function EventView() {
           <Text>Data e horário</Text>
           <Text>Descrição do evento</Text>
           <HStack>
-            <CustomTag texto="tag" visual="solid" />
-            <CustomTag texto="tag" visual="solid" />
-            <CustomTag texto="tag" visual="solid" />
-            <CustomTag texto="tag" visual="solid" />
+            <Tag label="tag" style="solid" />
+            <Tag label="tag" style="solid" />
+            <Tag label="tag" style="solid" />
+            <Tag label="tag" style="solid" />
           </HStack>
           <PrimaryCustomButton value="Editar evento" variant="primary" />
         </VStack>
