+++ src/pages/HomePage.tsx
@@ -3,7 +3,7 @@ import { Flex, Heading, IconButton } from "@chakra-ui/react";
 import { Plus } from "@phosphor-icons/react";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import CustomTag from "../components/buttons/CustomTag/CustomTag";
+import Tag from "../components/buttons/tag/tag";
 import EventCard from "../components/cards/EventCard";
 import SearchBar from "../components/layout/SearchBar/SearchBar";
 import { formatarData } from "../utils/formatData";
@@ -80,9 +80,7 @@ function HomePage() {
         <SearchBar placeHolder="Buscar evento" setTermoBusca={setTermoBusca} />
         <Flex gap={2} wrap="wrap">
           {tags.map((e, index) => {
-            return (
-              <CustomTag key={index} texto={`#${e.nome}`} visual={e.visual} />
-            );
+            return <Tag key={index} label={`#${e.nome}`} style={e.visual} />;
           })}
         </Flex>
         <Heading textStyle="2xl">Próximos Eventos</Heading>
